Added a conditional for both legs - baseline needs to be complete for both legs before the "complete" command is sent to iOS to request torque values. Tested.
In pid.ino, suddenly experienced compile error on leg->pid.Compute_KF. Changed to leg->pid.PID::Compute and successfully compiled. Not sure why compile error popped up now instead of 3+ years ago when it was written, or in the last few months when we started working on Nano, but alright.